+++ modules/aws/src/smithy4s/aws/AwsClient.scala
@@ -49,9 +49,7 @@ object AwsClient {
       awsService <- service.hints
         .get(_root_.aws.api.Service)
         .toRight(AwsClientInitialisationError.NotAws(service.id))
-      endpointPrefix <- awsService.endpointPrefix.toRight(
-        AwsClientInitialisationError.NoEndpointPrefix(awsService)
-      )
+      endpointPrefix = awsService.endpointPrefix.getOrElse(service.id.name)
       awsProtocol <- AwsProtocol(service.hints).toRight(
         AwsClientInitialisationError.UnsupportedProtocol(
           serviceId = service.id,
