+++ modules/compliance-tests/src/smithy4s/compliancetests/internals/DefaultSchemaVisitor.scala
@@ -109,6 +109,8 @@ private[compliancetests] object DefaultSchemaVisitor extends SchemaVisitor[Id] {
       refinement: Refinement[A, B]
   ): Id[B] = refinement.unsafe(apply(schema))
 
-  override def lazily[A](suspend: Lazy[Schema[A]]): Id[A] = ???
+  override def lazily[A](suspend: Lazy[Schema[A]]): Id[A] = {
+    suspend.map(apply).value
+  }
 
 }
